chore(DataTable): adding toggleColumn hook
